% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neightbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{Class value of training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
List with objects,
\code{class}: Vector of predicted class for all observations,
\code{cv_err}: Numeric with the cross-validation misclassification error.
}
\description{
This function predicts an output class using given covariates.
}
\examples{
data(my_penguins)
penguins_df <- tidyr::drop_na(my_penguins)
my_knn_cv(penguins_df[ , 3:6], penguins_df$species, 5, 5)

}
\keyword{prediction}
