% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object.}

\item{data}{input data frame.}
}
\value{
Data frame with rows for each coefficient and columns,
\code{Estimate}: a column with the solved linear regression coefficients,
\code{Standard_Error}: a column with the solved standard errors,
\code{t_value}: a column with the solved t-values,
\code{twoSided_test}: a column demonstrating a two-sided t-test for the covariates.
}
\description{
This function fits a linear model.
}
\examples{
data("mtcars")
my_lm(formula = mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
\keyword{prediction}
