% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric with number of folds.}
}
\value{
Numeric with cross-validation error
}
\description{
This Function performs Random Forest Cross-Validation.
}
\examples{
my_rf_cv(k = 5)

}
\keyword{prediction}
